a sketch of basic Worktree support (#301)
